назад

+++ backend/bot/handlers.py
@@ -9,7 +9,7 @@ from realties.models import Ad, Category, City, Comment, Favorite, Realty
 from users.models import Profile
 
 from .permissions import restricted
-from .utils import (chunks, get_botmessage_by_keyword, paginate,
+from .utils import (chunks, get_botmessage_by_keyword, paginate, split_query,
                     text_ad, text_realty)
 
 START, CITY, CITY_CHOICE, CATEGORY, PRICE = range(5)
@@ -321,7 +321,7 @@ async def next_page(update: Update, context: CallbackContext) -> int:
         if items.has_next():
             context.user_data['page'] = items.next_page_number()
             await update.message.reply_text(
-                '«Далее» для просмотра следующих объявлений.\n'
+                '«Далее» для просмотра следующих объявлений '
                 f'Вы посмотрели {items.end_index()} '
                 f'из {ad_queryset.count()}',
                 reply_markup=ReplyKeyboardMarkup(
@@ -359,7 +359,7 @@ async def next_page(update: Update, context: CallbackContext) -> int:
             if items.has_next():
                 context.user_data['page'] = items.next_page_number()
                 await update.message.reply_text(
-                    '«Далее» для просмотра следующих объектов недвижимости.\n'
+                    '«Далее» для просмотра следующих объектов недвижимости. '
                     f'Вы посмотрели {items.end_index()} '
                     f'из {realty_queryset.count()}',
                     reply_markup=ReplyKeyboardMarkup(
@@ -468,7 +468,8 @@ async def favorite(update: Update, context: CallbackContext):
                 InlineKeyboardButton(
                     "Удалить из избранного",
                     callback_data=f'{str(DELETE_FAVORITE)},'
-                                  f'{favorite_ad.pk}'
+                                  f'{favorite_ad.pk},'
+                                  f'{user_id}'
                 ),
                 InlineKeyboardButton(
                     "Комментарии",
@@ -486,17 +487,15 @@ async def favorite(update: Update, context: CallbackContext):
 
 async def delete_favorite(update: Update, context: CallbackContext):
     """Убираем объявление из избранного."""
-    user_id = update.effective_user.id
-    query = update.callback_query
-    ad_id = query.data.split(',')[1]
-    # if len(query_data) != 2:
-    #     await update.callback_query.edit_message_text(
-    #         'Некорректные данные для удаления.'
-    #     )
-    #     return
+    query_data = split_query(update)
+    if len(query_data) != 3:
+        await update.callback_query.edit_message_text(
+            'Некорректные данные для удаления.'
+        )
+        return
     try:
         Favorite.objects.get(
-            user__external_id=user_id, ad__pk=ad_id
+            user__external_id=query_data[2], ad__pk=query_data[1]
         ).delete()
         await update.callback_query.edit_message_text(
             'Запись удалена из избранного!'
